+++ ui/src/shared/components/Dygraph.js
@@ -9,6 +9,7 @@ import getRange from 'shared/parsing/getRangeForDygraph'
 
 import {LINE_COLORS, multiColumnBarPlotter} from 'src/shared/graphs/helpers'
 import DygraphLegend from 'src/shared/components/DygraphLegend'
+import {buildYLabel} from 'shared/presenters'
 
 export default class Dygraph extends Component {
   constructor(props) {
@@ -35,12 +36,14 @@ export default class Dygraph extends Component {
     this.handleHideLegend = ::this.handleHideLegend
     this.handleToggleFilter = ::this.handleToggleFilter
     this.visibility = ::this.visibility
+    this.getLabel = ::this.getLabel
   }
 
   static defaultProps = {
     containerStyle: {},
     isGraphFilled: true,
     overrideLineColors: null,
+    dygraphRef: () => {},
   }
 
   getTimeSeries() {
@@ -50,11 +53,23 @@ export default class Dygraph extends Component {
     return timeSeries.length ? timeSeries : [[0]]
   }
 
+  getLabel(axis) {
+    const {axes, queries} = this.props
+    const label = _.get(axes, [axis, 'label'], '')
+    const queryConfig = _.get(queries, ['0', 'queryConfig'], false)
+
+    if (label || !queryConfig) {
+      return label
+    }
+
+    return buildYLabel(queryConfig)
+  }
+
   componentDidMount() {
     const timeSeries = this.getTimeSeries()
     // dygraphSeries is a legend label and its corresponding y-axis e.g. {legendLabel1: 'y', legendLabel2: 'y2'};
     const {
-      ranges,
+      axes,
       dygraphSeries,
       ruleValues,
       overrideLineColors,
@@ -70,6 +85,9 @@ export default class Dygraph extends Component {
     if (finalLineColors === null) {
       finalLineColors = LINE_COLORS
     }
+
+    const yAxis = _.get(axes, ['y', 'bounds'], [null, null])
+    const y2Axis = _.get(axes, ['y2', 'bounds'], undefined)
 
     const defaultOptions = {
       plugins: [
@@ -90,12 +108,13 @@ export default class Dygraph extends Component {
       hideOverlayOnMouseOut: false,
       colors: finalLineColors,
       series: dygraphSeries,
+      ylabel: this.getLabel('y'),
       axes: {
         y: {
-          valueRange: getRange(timeSeries, ranges.y, ruleValues),
+          valueRange: getRange(timeSeries, yAxis, ruleValues),
         },
         y2: {
-          valueRange: getRange(timeSeries, ranges.y2),
+          valueRange: getRange(timeSeries, y2Axis),
         },
       },
       highlightSeriesOpts: {
@@ -235,7 +254,7 @@ export default class Dygraph extends Component {
   componentDidUpdate() {
     const {
       labels,
-      ranges,
+      axes,
       options,
       dygraphSeries,
       ruleValues,
@@ -249,16 +268,21 @@ export default class Dygraph extends Component {
       )
     }
 
+    const y = _.get(axes, ['y', 'bounds'], [null, null])
+    const y2 = _.get(axes, ['y2', 'bounds'], undefined)
     const timeSeries = this.getTimeSeries()
+    const ylabel = this.getLabel('y')
+
     const updateOptions = {
       labels,
       file: timeSeries,
+      ylabel,
       axes: {
         y: {
-          valueRange: getRange(timeSeries, ranges.y, ruleValues),
+          valueRange: getRange(timeSeries, y, ruleValues),
         },
         y2: {
-          valueRange: getRange(timeSeries, ranges.y2),
+          valueRange: getRange(timeSeries, y2),
         },
       },
       stepPlot: options.stepPlot,
@@ -350,6 +374,7 @@ export default class Dygraph extends Component {
         <div
           ref={r => {
             this.graphRef = r
+            this.props.dygraphRef(r)
           }}
           style={this.props.containerStyle}
           className="dygraph-child-container"
@@ -359,13 +384,18 @@ export default class Dygraph extends Component {
   }
 }
 
-const {array, arrayOf, func, number, bool, shape, string} = PropTypes
+const {array, arrayOf, bool, func, shape, string} = PropTypes
 
 Dygraph.propTypes = {
-  ranges: shape({
-    y: arrayOf(number),
-    y2: arrayOf(number),
+  axes: shape({
+    y: shape({
+      bounds: array,
+    }),
+    y2: shape({
+      bounds: array,
+    }),
   }),
+  queries: arrayOf(shape),
   timeSeries: array.isRequired,
   labels: array.isRequired,
   options: shape({}),
@@ -384,4 +414,5 @@ Dygraph.propTypes = {
   }),
   synchronizer: func,
   setResolution: func,
+  dygraphRef: func,
 }